Rewrote roller velocity timer, made it worse

+++ src/drivers/inc_encoder.h
@@ -17,6 +17,6 @@
 void inc_encoder_init();
 float inc_encoder_get_pos();
 float inc_encoder_get_vel();
-uint32_t inc_encoder_get_dt_ticks();
+uint16_t inc_encoder_get_dt_ticks();
 
 #endif /* INC_ENCODER_H_ */
